create camera container
